requirements.txt
absl-py==1.4.0
albumentations==1.3.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Flask==2.2.2
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-cloud-vision==3.3.1
google-pasta==0.2.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h5py==3.8.0
idna==3.4
imageio==2.25.0
importlib-metadata==6.0.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pathlib==1.0.1
Pillow==9.4.0
proto-plus==1.22.2
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-version==0.0.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sklearn==0.0.post2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.1.23.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
